Allow adding empty functions at specific addresses
Enables adding a function that is a single instruction of the undefined
intermediate instruction at a specified address. This is useful when you
want to completely emulate a function without it actually being added to
the program.

A user can just hook the entry of the function, emulate the function
behavior in the hook, and then jump to the return address.