Correctly order command line arguments
The command line arguments were incorrectly being passed to the program
in reverse order. They are pushed onto a stack in order and then passed
to the program in the order of the stack. The stack needs to be reversed
to maintain the order of the arguments.

Adds default page size of 0x1000 that will be passed to programs.